+++ scripts/macros/2024/spells/dream.js
@@ -1,0 +1,5 @@
+export let dream = {
+    name: 'Dream',
+    version: '1.3.160',
+    rules: 'modern'
+};